+++ src/main/java/strategies/macdOverRSIStrategies/Short/MACDOverRSIShortExitStrategy5.java
@@ -22,7 +22,7 @@ public class MACDOverRSIShortExitStrategy5 extends MACDOverRSIBaseExitStrategy {
 
     @Override
     public SellingInstructions run(DataHolder realTimeData) {
-        BigDecimal currentPrice = realTimeData.getCurrentPrice();
+        double currentPrice = realTimeData.getCurrentPrice();
         if (! isTrailing){
             trailer.setAbsoluteMaxPrice(currentPrice);
             isTrailing = true;
